BIG REFACTOR!! (replace ui module with canvas)
Yeah, this is the second refactor of this thing... But now I'm really
happy with it. Instead of Ui and Widgets, it's Canvas and CanvasMode
now. Not much of a reason for that naming change, but I also changed
some of the interfaces, and it feels a lot better now